chore: update from int to long

+++ src/EulerMaths/Problem1.cs
@@ -8,7 +8,7 @@ namespace EulerMaths
   /// </summary>
   public static class Problem1
   {
-    public static int Answer()
+    public static long Answer()
     {
       return Enumerable.Range(1, 1000).Where(n => n % 3 == 0 || n % 5 == 0).Sum();
     }
